Replace integer with long

+++ src/main/java/com/thekkid/ninetyninelambdas/problems/Twenty.java
@@ -14,8 +14,8 @@ import static com.jnape.palatable.lambda.functions.builtin.fn2.Map.map;
  * Remove the K'th element from a list
  */
 public class Twenty {
-    public static <T> Iterable<T> twenty(int n, Iterable<T> ts) {
-        return Zip.<Integer, T>zip(iterate(fn2(Integer::sum).apply(1), 1))
+    public static <T> Iterable<T> twenty(long n, Iterable<T> ts) {
+        return Zip.<Long, T>zip(iterate(fn2(Long::sum).apply(1L), 1L))
                 .fmap(filter(not(cmpEq(n).contraMap(Tuple2::_1))))
                 .fmap(map(Tuple2::_2))
                 .apply(ts);
